refactor(*): Lotto, WinningNumbers data package로 이동
Lotto, WinningNumbers data package로 이동시켰습니다.
