+++ src/main/java/lotto/domain/LottoPurchaseManager.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.List;
+import lotto.data.Lotto;
 import lotto.message.LottoResult;
 import lotto.message.OutputMessage;
 import lotto.utils.LottoUtil;
